fixed redirect loop issue due to load balancer round-robin
problem: On the backend, my blog's server listens on two ports, one for HTTP and the other for HTTPS. The HTTP port just sends a redirect to HTTPS. Now, I don't need this anymore because my load balancer is terminating the SSL connection. But, since I have both those ports exposed, it's doing a round-robin to both of them. So you have a 50/50 chance of getting redirected to the same location anytime you connect, based on whether the load balancer chooses the HTTP port or the HTTPS port.